Split x86 mod
